+++ ezdxf/ac1009/graphics.py
@@ -955,3 +955,40 @@ class Dimension(GraphicEntity):
         'oblique_angle': DXFAttr(52),
         'rotation_angle': DXFAttr(53),
     })
+
+_SHAPE_TPL = """  0
+SHAPE
+  5
+0
+  8
+0
+ 10
+0.0
+ 20
+0.0
+ 30
+0.0
+ 40
+1.0
+  2
+NAME
+ 50
+0.0
+ 41
+1.0
+ 51
+0.0
+"""
+
+
+# SHAPE is not tested with real world DXF drawings!
+class Shape(GraphicEntity, ColorMixin):
+    TEMPLATE = ClassifiedTags.from_text(_SHAPE_TPL)
+    DXFATTRIBS = make_attribs({
+        'insert': DXFAttr(10, xtype='Point2D/3D'),
+        'size': DXFAttr(40),
+        'name': DXFAttr(2),
+        'rotation': DXFAttr(50, default=0.0),
+        'xscale': DXFAttr(41, default=1.0),
+        'oblique': DXFAttr(51, default=0.0),
+    })