+++ NEWS.rst
@@ -9,6 +9,7 @@ Version 0.5.0 - 2014-04-..
   * added default values to DXF attributes as defined in the DXF standard, this allows to check optional DXF attributes
     (with defined default values) to be used without check of presence, like *entity.dxf.paperspace*.
   * entity 3DFACE requires 3D coordinates (created by add_3Dface())
+  * new supported DXF entities: Shape, Ray, XLine
 
 Version 0.4.2 - 2014-04-02